gralloc: Define custom usages in gralloc_priv.h
This removes the need for a custom gralloc header in device trees with
the custom usages defined within it. Having the custom gralloc header
somehow causes Codec2 to not work and, when VNDK is enabled, ABI tests
for it to fail.

Also, only define them if they're not already defined in an overridden
gralloc.h header for existing device trees which don't have it removed.

Change-Id: I296be2484aed6faafee741ea1f68b067edc7a859